+++ server/src/main/scala/com/stratio/crossdata/server/actors/ServerActor.scala
@@ -15,54 +15,57 @@
  */
 package com.stratio.crossdata.server.actors
 
+import java.io.{File, InputStream}
+import java.lang.reflect.Method
+import java.net.{URL, URLClassLoader}
 import java.util.UUID
 
-import akka.cluster.ClusterEvent.InitialStateAsSnapshot
-import akka.contrib.pattern.DistributedPubSubExtension
 import akka.actor.SupervisorStrategy.Restart
 import akka.actor._
 import akka.cluster.Cluster
-import akka.contrib.pattern.DistributedPubSubMediator.{Publish, SubscribeAck, Subscribe}
+import akka.cluster.ClusterEvent.InitialStateAsSnapshot
+import akka.contrib.pattern.DistributedPubSubExtension
+import akka.contrib.pattern.DistributedPubSubMediator.{Publish, Subscribe, SubscribeAck}
 import akka.remote.DisassociatedEvent
-import com.stratio.crossdata.common.result.{ErrorSQLResult, SuccessfulSQLResult}
+import com.google.common.io.Files
 import com.stratio.crossdata.common.security.Session
-import com.stratio.crossdata.common._
-import com.stratio.crossdata.server.actors.JobActor.Commands.{StartJob, CancelJob}
-import com.stratio.crossdata.common.{CommandEnvelope, SQLCommand}
+import com.stratio.crossdata.common.{CommandEnvelope, SQLCommand, _}
+import com.stratio.crossdata.server.actors.JobActor.Commands.{CancelJob, StartJob}
 import com.stratio.crossdata.server.actors.JobActor.Events.{JobCompleted, JobFailed}
 import com.stratio.crossdata.server.config.{ServerActorConfig, ServerConfig}
 import org.apache.log4j.Logger
 import org.apache.spark.sql.crossdata.XDContext
-import org.apache.spark.sql.types.StructType
 
 import scala.concurrent.duration.FiniteDuration
-import scala.reflect.io.File
 
 
 object ServerActor {
-
-  val managementTopic: String = "jobsManagement"
+  val ManagementTopic: String = "jobsManagement"
 
   def props(cluster: Cluster, xdContext: XDContext, config: ServerActorConfig): Props =
     Props(new ServerActor(cluster, xdContext, config))
+
 
   protected case class JobId(requester: ActorRef, queryId: UUID)
 
   protected case class ManagementEnvelope(command: ControlCommand, source: ActorRef)
 
   private object ManagementMessages {
+
     case class DelegateCommand(scommand: CommandEnvelope, broadcaster: ActorRef)
+
     case class FinishJob(jobActor: ActorRef)
+
   }
 
   case class State(jobsById: Map[JobId, ActorRef])
 
 }
 
-class ServerActor(cluster: Cluster, xdContext: XDContext, config: ServerActorConfig) extends Actor {
+class ServerActor(cluster: Cluster, xdContext: XDContext, serverActorConfig: ServerActorConfig) extends Actor with ServerConfig {
 
-  import ServerActor._
   import ServerActor.ManagementMessages._
+  import ServerActor._
 
   lazy val logger = Logger.getLogger(classOf[ServerActor])
 
@@ -72,7 +75,7 @@ class ServerActor(cluster: Cluster, xdContext: XDContext, config: ServerActorCon
     super.preStart()
 
     // Subscribe to the management distributed topic
-    mediator ! Subscribe(managementTopic, self)
+    mediator ! Subscribe(ManagementTopic, self)
 
     // Subscribe to disassociation events in the cluster
     cluster.subscribe(self, InitialStateAsSnapshot, classOf[DisassociatedEvent])
@@ -81,12 +84,19 @@ class ServerActor(cluster: Cluster, xdContext: XDContext, config: ServerActorCon
 
   // Actor behaviours
 
-  override def receive: Actor.Receive = initial
+  override def receive: Actor.Receive = initial(Set(ManagementTopic))
 
-  private val initial: Receive = {
-    case SubscribeAck(Subscribe(managementTopic, None, self)) =>
-      context.become(ready(State(Map.empty)))
+  private def initial(pendingTopics: Set[String]): Receive = {
+    case SubscribeAck(Subscribe(ManagementTopic, None, self)) =>
+      val newPendingTopics = pendingTopics - ManagementTopic
+      checkSubscriptions(newPendingTopics)
   }
+
+  private def checkSubscriptions(pendingTopics: Set[String]): Unit =
+    if (pendingTopics.isEmpty)
+      context.become(ready(State(Map.empty)))
+    else
+      context.become(initial(pendingTopics))
 
   /**
     * If a `cmd` is passed to this method is because it has already been checked that this server can run it.
@@ -95,26 +105,36 @@ class ServerActor(cluster: Cluster, xdContext: XDContext, config: ServerActorCon
     * @param st
     */
   private def executeAccepted(cmd: CommandEnvelope)(st: State): Unit = cmd match {
-    case CommandEnvelope(sqlCommand @ SQLCommand(query, queryId, withColnames, timeout), session @ Session(id, requester)) =>
+    case CommandEnvelope(sqlCommand@SQLCommand(query, queryId, withColnames, timeout), session@Session(id, requester)) =>
       logger.debug(s"Query received $queryId: $query. Actor ${self.path.toStringWithoutAddress}")
       logger.debug(s"Session identifier $session")
       val jobActor = context.actorOf(JobActor.props(xdContext, sqlCommand, sender(), timeout))
       jobActor ! StartJob
       context.become(ready(st.copy(jobsById = st.jobsById + (JobId(requester, sqlCommand.queryId) -> jobActor))))
 
-    case CommandEnvelope(addJarCommand:AddJARCommand,session @ Session(id, requester)) =>
-      logger.debug(s"Add JAR received ${addJarCommand.requestId}: ${addJarCommand.path}. Actor ${self.path.toStringWithoutAddress}")
-      logger.debug(s"Session identifier $session")
-      //TODO  Maybe include job controller if it is necessary as in sql command
-      if (addJarCommand.path.toLowerCase.startsWith("hdfs://") || File(addJarCommand.path).exists) {
-        xdContext.addJar(addJarCommand.path)
-        sender ! SQLReply(addJarCommand.requestId, SuccessfulSQLResult(Array.empty, new StructType()))
-      } else {
-        sender ! SQLReply(addJarCommand.requestId, ErrorSQLResult("File doesn't exists or is not a hdfs file", Some(new Exception("File doesn't exists or is not a hdfs file"))))
-      }
-
-    case CommandEnvelope(cc @ CancelQueryExecution(queryId), session @ Session(id, requester)) =>
+    case CommandEnvelope(cc@CancelQueryExecution(queryId), session@Session(id, requester)) =>
       st.jobsById.get(JobId(requester, queryId)).get ! CancelJob
+  }
+
+  private def addToClasspath(file: File): Unit = {
+    if (file.exists) {
+      val method: Method = classOf[URLClassLoader].getDeclaredMethod("addURL", classOf[URL])
+      method.setAccessible(true)
+      method.invoke(ClassLoader.getSystemClassLoader, file.toURI.toURL)
+      method.setAccessible(false)
+    } else {
+      logger.warn(s"The file ${file.getName} not exists.")
+    }
+  }
+
+  private def createFile(hdfsIS: InputStream, path: String): File = {
+    val targetFile = new File(path)
+
+    val arrayBuffer = new Array[Byte](hdfsIS.available)
+    hdfsIS.read(arrayBuffer)
+
+    Files.write(arrayBuffer, targetFile)
+    targetFile
   }
 
   // Receive functions:
@@ -124,9 +144,10 @@ class ServerActor(cluster: Cluster, xdContext: XDContext, config: ServerActorCon
     case DelegateCommand(_, broadcaster) if broadcaster == self => //Discards from this server broadcast delegated-commands
 
     case DelegateCommand(cmd, broadcaster) if broadcaster != self =>
-      cmd match { // Inner pattern matching for future delegated command validations
-        case sc @ CommandEnvelope(CancelQueryExecution(queryId), Session(_, requester)) =>
-          st.jobsById.get(JobId(requester, queryId)) foreach(_ => executeAccepted(sc)(st))
+      cmd match {
+        // Inner pattern matching for future delegated command validations
+        case sc@CommandEnvelope(CancelQueryExecution(queryId), Session(_, requester)) =>
+          st.jobsById.get(JobId(requester, queryId)) foreach (_ => executeAccepted(sc)(st))
         /* If it doesn't validate it won't be re-broadcast since the source server already distributed it to all
             servers through the topic. */
       }
@@ -135,19 +156,24 @@ class ServerActor(cluster: Cluster, xdContext: XDContext, config: ServerActorCon
   // Commands reception: Checks whether the command can be run at this Server passing it to the execution method if so
   def commandMessagesRec(st: State): Receive = {
 
-    case sc @ CommandEnvelope(_: SQLCommand, _) =>
+    case sc@CommandEnvelope(_: SQLCommand, _) =>
       executeAccepted(sc)(st)
 
-    case sc @ CommandEnvelope(_: AddJARCommand, _) =>
+    // TODO broadcastMessage
+    case sc@CommandEnvelope(_: AddJARCommand, _) =>
       executeAccepted(sc)(st)
 
-    case sc @ CommandEnvelope(cc: ControlCommand, session @ Session(id, requester)) =>
+    case sc@CommandEnvelope(cc: ControlCommand, session@Session(id, requester)) =>
       st.jobsById.get(JobId(requester, cc.requestId)) map { _ =>
         executeAccepted(sc)(st) // Command validated to be executed by this server.
       } getOrElse {
         // If it can't run here it should be executed somewhere else
-        mediator ! Publish(managementTopic, DelegateCommand(sc, self))
+        mediator ! Publish(ManagementTopic, DelegateCommand(sc, self))
       }
+
+    case clusterStateCommand @ ClusterStateCommand() =>
+      sender ! ClusterStateReply(clusterStateCommand.requestId, cluster.state)
+
 
   }
 
@@ -172,27 +198,27 @@ class ServerActor(cluster: Cluster, xdContext: XDContext, config: ServerActorCon
       commandMessagesRec(st) orElse
       eventsRec(st) orElse {
 
-    case DisassociatedEvent(_, remoteAddress, _) =>
-      /*
-        Uses `DisassociatedEvent` to get notified of an association loss.
-        An akka association consist on the connection of a JVM to another to build remote connections upon.
-        Thus, the reception of this event message means all remote clients within the addressed jvm are down.
+      case DisassociatedEvent(_, remoteAddress, _) =>
+        /*
+          Uses `DisassociatedEvent` to get notified of an association loss.
+          An akka association consist on the connection of a JVM to another to build remote connections upon.
+          Thus, the reception of this event message means all remote clients within the addressed jvm are down.
 
-         More info at: http://doc.akka.io/docs/akka/2.3.11/scala/remoting.html
-       */
-      val newjobsmap = st.jobsById filter {
-        case (JobId(requester, _), job) if requester.path.address == remoteAddress =>
-          gracefullyKill(job) // WARNING! Side-effect
-          false
-        case _ => true
-      }
-      context.become(ready(st.copy(jobsById = newjobsmap)))
+           More info at: http://doc.akka.io/docs/akka/2.3.11/scala/remoting.html
+         */
+        val newjobsmap = st.jobsById filter {
+          case (JobId(requester, _), job) if requester.path.address == remoteAddress =>
+            gracefullyKill(job) // WARNING! Side-effect
+            false
+          case _ => true
+        }
+        context.become(ready(st.copy(jobsById = newjobsmap)))
 
-    case any =>
-      logger.warn(s"Something is going wrong! Unknown message: $any")
-  }
+      case any =>
+        logger.warn(s"Something is going wrong! Unknown message: $any")
+    }
 
-  private def sentenceToDeath(victim: ActorRef): Unit = config.completedJobTTL match {
+  private def sentenceToDeath(victim: ActorRef): Unit = serverActorConfig.completedJobTTL match {
     case finite: FiniteDuration =>
       context.system.scheduler.scheduleOnce(finite, self, FinishJob(victim))(context.dispatcher)
     case _ => // Reprieve by infinite limit
@@ -204,7 +230,7 @@ class ServerActor(cluster: Cluster, xdContext: XDContext, config: ServerActorCon
   }
 
   //TODO: Use number of tries and timeout configuration parameters
-  override def supervisorStrategy: SupervisorStrategy = OneForOneStrategy(config.retryNoAttempts, config.retryCountWindow) {
+  override def supervisorStrategy: SupervisorStrategy = OneForOneStrategy(serverActorConfig.retryNoAttempts, serverActorConfig.retryCountWindow) {
     case _ => Restart //Crashed job gets restarted (or not, depending on `retryNoAttempts` and `retryCountWindow`)
   }