+++ testsIT/src/test/scala/com/stratio/crossdata/driver/DriverIT.scala
@@ -21,9 +21,9 @@ import com.stratio.crossdata.common.result.{ErrorSQLResult, SuccessfulSQLResult}
 import org.junit.runner.RunWith
 import org.scalatest.junit.JUnitRunner
 
+import scala.concurrent.Await
 import scala.concurrent.duration._
 import scala.language.postfixOps
-import scala.reflect.io.File
 
 @RunWith(classOf[JUnitRunner])
 class DriverIT extends EndToEndTest {
@@ -59,6 +59,7 @@ class DriverIT extends EndToEndTest {
   }
 
   it should "get a list of tables" in {
+    assumeCrossdataUpAndRunning
     val driver = Driver.getOrCreate()
 
     driver.sql(
@@ -73,7 +74,9 @@ class DriverIT extends EndToEndTest {
   }
 
   "Crossdata Driver" should "be able to close the connection and start it again" in {
-    var driver = Driver.getOrCreate(); Driver.getOrCreate()
+    assumeCrossdataUpAndRunning
+    var driver = Driver.getOrCreate();
+    Driver.getOrCreate()
     val newDriver = Driver.getOrCreate()
 
     driver should be theSameInstanceAs newDriver
@@ -89,38 +92,67 @@ class DriverIT extends EndToEndTest {
 
     val result = driver.sql(s"SHOW TABLES")
 
-    result.hasError should equal (false)
+    result.hasError should equal(false)
 
   }
 
 
-  it should "be able to execute ADD JAR Command of an existent file" in {
-    val file=File("/tmp/jar").createFile(false)
+  //    // TODO move to examples??
+  //    // TODO check this tests with HDFS
+  //    it should "be able to execute ADD JAR Command of an existent file" in {
+  //      assumeCrossdataUpAndRunning
+  //
+  //      val file=File(s"/tmp/bulk_${System.currentTimeMillis()}.jar").createFile(false)
+  //      val driver = Driver.getOrCreate()
+  //      val result = driver.addJar(file.path).waitForResult()
+  //
+  //      driver.stop()
+  //      file.delete()
+  //
+  //      result.hasError should equal (false)
+  //    }
+
+  //  it should "be return an Error when execute ADD JAR Command of an un-existent file" in {
+  //
+  //    val driver = Driver.getOrCreate()
+  //    val result = driver.addJar(s"/tmp/jarnotexists").waitForResult()
+  //    driver.stop()
+  //
+  //    result.hasError should equal (true)
+  //  }
+  //
+  //  it should "be able to execute ADD JAR Command of any HDFS file" in {
+  //
+  //    val driver = Driver.getOrCreate()
+  //    val result = driver.addJar(s"hdfs://repo/file.jar").waitForResult()
+  //    driver.stop()
+  //
+  //    result.hasError should equal (false)
+  //  }
+
+
+  it should "indicates that the cluster is alive when there is a server up" in {
     val driver = Driver.getOrCreate()
-    val result = driver.addJar(s"/tmp/jar").waitForResult()
 
-    driver.stop()
-    file.delete()
-
-    result.hasError should equal (false)
+    driver.isClusterAlive(6 seconds) shouldBe true
   }
 
-  it should "be return an Error when execute ADD JAR Command of an un-existent file" in {
-
+  it should "return the addresses of servers up and running" in {
     val driver = Driver.getOrCreate()
-    val result = driver.addJar(s"/tmp/jarnotexists").waitForResult()
-    driver.stop()
 
-    result.hasError should equal (true)
+    val addresses = Await.result(driver.serversUp(), 6 seconds)
+
+    addresses should have length 1
+    addresses.head.host shouldBe Some("127.0.0.1")
   }
 
-  it should "be able to execute ADD JAR Command of any HDFS file" in {
-
+  it should "return the current cluster state" in {
     val driver = Driver.getOrCreate()
-    val result = driver.addJar(s"hdfs://repo/file.jar").waitForResult()
-    driver.stop()
 
-    result.hasError should equal (false)
+    val clusterState = Await.result(driver.clusterState(), 6 seconds)
+
+    clusterState.getLeader.host shouldBe Some("127.0.0.1")
   }
+
 
 }