+++ CHANGELOG.md
@@ -3,7 +3,25 @@
 Only listing significant user-visible, not internal code cleanups and minor bug fixes. 
 
 ## 1.3.0 (upcoming)
+* Upgrade to Spark 1.6.1
+* Support for INSERT INTO
+* Support for ADD JAR
 * Enabled Scala cross builds for Scala 2.10 and Scala 2.11.
+* Encryption for the communication between client and server.
+* Added method to check communication between client ans server.
+* Upgrade Stratio spark-mongodb to 0.11.2
+* Improved communication between client and server.
+* Native jobs available in SparkUI
+* Improved native access to MongoDB
+
+## 1.2.2 (May 2016)
+* Upgrade spark-mongodb to 0.11.2
+* Security Manager added
+* Deploying uber-jars in Maven
+
+## 1.2.1 (April 2016)
+* Deploy mode parameter for Spark-Submit configuration added
+* Bux fixing in package creation
 
 ## 1.2.0 (March 2016)
 * StreamSQL: Streaming capabilities from a SQL language including insert into select (Batch + Streaming)